+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2024-08-03
+
+### Added
+- Added cspReportRoute Endpoint for CSP violations
+- Added debug level to winston logger for dev enviroment
+
+### Changed
+- Changed ContentSecurityPolicy Headers directive for improved security
+
 ## [v16.12.4] - 2024-08-01
 
 ### Fixed